+++ tests/testthat/test-prop_diff.R
@@ -8,7 +8,7 @@ testthat::test_that("`prop_diff_ha` (proportion difference by Anderson-Hauck)", 
   # according to SAS.
   expected <- list(
     diff = 0.25,
-    diff_ci = c(-0.9195, 1)
+    diff_ci = c(-1.0000, 0.9195)
   )
   testthat::expect_equal(result, expected, tol = 0.0001)
 
@@ -38,20 +38,72 @@ testthat::test_that("`prop_diff_nc` (proportion difference by Newcombe)", {
   # according to SAS.
   expected <- list(
     diff = 0.25,
-    diff_ci = c(-0.2967, 0.6750)
+    diff_ci = c(-0.6750, 0.2967)
   )
   testthat::expect_equal(result, expected, tol = 0.0001)
 
   # Edge case: Same proportion of response in A and B.
   rsp <- c(TRUE, FALSE, TRUE, FALSE)
   grp <- factor(c("A", "A", "B", "B"), levels = c("A", "B"))
-  result <- testthat::expect_warning(
-    prop_diff_nc(rsp = rsp, grp = grp, conf_level = 0.6)
-  )
+  result <- prop_diff_nc(rsp = rsp, grp = grp, conf_level = 0.6)
+
   # according to SAS.
   expected <- list(
     diff = 0,
     diff_ci = c(-0.3616, 0.3616)
+  )
+  testthat::expect_equal(result, expected, tol = 0.0001)
+})
+
+testthat::test_that("`prop_diff_wald` (proportion difference by Wald's test: with correction)", {
+
+  # "Mid" case: 3/4 respond in group A, 1/2 respond in group B.
+  rsp <- c(TRUE, FALSE, FALSE, TRUE, TRUE, TRUE)
+  grp <- factor(c("A", "B", "A", "B", "A", "A"), levels = c("B", "A"))
+
+  result <- prop_diff_wald(rsp = rsp, grp = grp, conf_level = 0.9, correct = TRUE)
+
+  expected <- list(
+    diff = 0.25,
+    diff_ci = c(-1.0000, 0.8069)
+  )
+  testthat::expect_equal(result, expected, tol = 0.0001)
+
+  # Edge case: Same proportion of response in A and B.
+  rsp <- c(TRUE, FALSE, TRUE, FALSE)
+  grp <- factor(c("A", "A", "B", "B"), levels = c("A", "B"))
+  result <- prop_diff_wald(rsp = rsp, grp = grp, conf_level = 0.6, correct = TRUE)
+
+  expected <- list(
+    diff = 0,
+    diff_ci = c(-0.9208, 0.9208)
+  )
+  testthat::expect_equal(result, expected, tol = 0.0001)
+})
+
+testthat::test_that("`prop_diff_wald` (proportion difference by Wald's test: without correction)", {
+
+  # "Mid" case: 3/4 respond in group A, 1/2 respond in group B.
+  rsp <- c(TRUE, FALSE, FALSE, TRUE, TRUE, TRUE)
+  grp <- factor(c("A", "B", "A", "B", "A", "A"), levels = c("B", "A"))
+
+  result <- suppressWarnings(
+    prop_diff_wald(rsp = rsp, grp = grp, conf_level = 0.9, correct = FALSE)
+  )
+  expected <- list(
+    diff = 0.25,
+    diff_ci = c(-0.9319, 0.4319)
+  )
+  testthat::expect_equal(result, expected, tol = 0.0001)
+
+  # Edge case: Same proportion of response in A and B.
+  rsp <- c(TRUE, FALSE, TRUE, FALSE)
+  grp <- factor(c("A", "A", "B", "B"), levels = c("A", "B"))
+  result <- prop_diff_wald(rsp = rsp, grp = grp, conf_level = 0.6, correct = FALSE)
+
+  expected <- list(
+    diff = 0,
+    diff_ci = c(-0.4208, 0.4208)
   )
   testthat::expect_equal(result, expected, tol = 0.0001)
 })
@@ -139,7 +191,7 @@ testthat::test_that("`estimate_proportion_diff` is compatible with `rtables`", {
   expected <- structure(
     c(
       "", "Difference in Response rate (%)", "90% CI (Anderson-Hauck)",
-      "B", "", "", "A", "25.0", "(-92.0, 100.0)"
+      "B", "", "", "A", "25.0", "(-100.0, 92.0)"
     ),
     .Dim = c(3L, 3L)
   )