+++ NEWS.md
@@ -1,6 +1,7 @@
 # tern 0.7.8.9025.1
 
 ### Enhancements 
+* Added `DescTools` `BinomDiffCI` function within `tern`.
 * exported `draw_grob` function (needed by `enableRF`).
 * `checkmate::assert()`, `checkmate::assert_true()`, and `checkmate::assert_false()`
    use is kept to a minimum due to its ambiguous messages. 
@@ -9,6 +10,7 @@
   its length or by using `checkmate::assert_vector()`.
 * Removed `assert_equal_length` as comparing lengths of multiple vectors can be 
   done without the need of a custom function.
+  to implement `Yates` correction term.
 * Implemented `nestcolor` in all examples with slight refactoring to `g_km`, `g_ipp`, 
   `g_waterfall`, `g_step`, `g_lineplot`, and `g_forest`.